test(react): merge the seam test into element.test
