+++ packages/react/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2020",
+    "module": "esnext",
     "useDefineForClassFields": true,
     "esModuleInterop": true,
     "jsx": "react-jsx",
